requirements.txt
cmat==0.1
einops==0.4.1
numpy==1.23.0
pandas==1.4.3
pytorch-lightning==1.6.4
PyYAML==6.0
ray==2.2.0
scikit-learn==1.1.1
scipy==1.8.1
tabulate==0.9.0
torch==1.12.0
torchmetrics==0.9.2
tqdm==4.64.0
wandb==0.12.21
zipp==3.8.0
